Move result generation to ProcessService
