+++ Kitodo/src/main/java/org/kitodo/forms/BaseForm.java
@@ -34,7 +34,6 @@ public class BaseForm implements Serializable {
     private boolean saveDisabled = true;
 
     // error messages
-    protected static final String ERROR_CREATING = "errorCreating";
     protected static final String ERROR_DATABASE_READING = "errorDatabaseReading";
     protected static final String ERROR_DUPLICATE = "errorDuplicate";
     protected static final String ERROR_EXPORTING = "errorExporting";
